+++ uui/src/lib/TextField/TextField.types.d.ts
@@ -9,8 +9,7 @@ export interface TextFieldProps extends svelte.JSX.HTMLAttributes<HTMLInputEleme
   value?: string;
   helperText?: string;
   color?: ThemeColors;
-  // Needs better name
-  onSurfaceColor?: ThemeColors;
+  error?: boolean;
   characterCounter?: boolean;
 }