+++ changelog.md
@@ -100,12 +100,12 @@
 
 - Created outlined and filled variants of TextField component
 - - Accepts all native HTMLInput properties as well as:
-- - - variant: "outlined" | "filled"
+- - - variant: "outlined" | "filled" = 'outlined' (default)
 - - - label: string
 - - - helperText?: string
 - - - characterCounter?: boolean (must be used with maxLength)
 - - - color?: ThemeColors = inherit (default)
-- - - onSurfaceColor?: ThemeColors = inherit (default) ...?
+- - - error?: boolean = false (default)
 - Created TextFieldIcon component specifically for TextField leading and trailing icons
 - - accepts leadingIcon and trailingIcon as slot names
 - - expects variant ("leading" | "trailing")